modificaciones productos y picadas
ademas de modificar y agregar clases como Combo, me fije algunas ideas para realizar la verificacion

+++ picada-mdq/src/Ingredientes/producto/Producto.java
@@ -1,9 +1,11 @@
 package Ingredientes.producto;
 //vamos a tener los tipos de queso,fiambre y snack, + el stock
+//Vamos a trabajar con enteros simulando que son bolsas de queso, bolsas de jamon y bolsas de snack
+
 public abstract class Producto {
     private int stock;
 
-    public Producto(double stock) {
+    public Producto(int stock) {
         this.stock = stock;
     }
 
@@ -11,12 +13,14 @@ public abstract class Producto {
         return stock;
     }
 
-    public void setStock(double stock) {
+    public void setStock(int stock) {
         this.stock = stock;
     }
 
     //funcion que modifica el stock en base a lo que pidio el usuario
     public abstract void actualizarStock();
+
+    //
 
 
 }
